pyproject.toml
[tool.poetry]
name = "virtool-workflow"
version = "0.0.0"
description = "A framework for developing bioinformatics workflows for Virtool."
authors = ["Ian Boyes", "Blake Smith", "Ryan Fang"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/virtool/virtool-workflow"
classifiers = [
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3.10"
]
packages = [
    { include = "virtool_workflow" },
]

[tool.poetry.dependencies]
python = "~3.10"
click = "^8.1.7"
aiohttp = "^3.8.1"
aiofiles = "^0.7.0"
virtool-core = "^8.0.0"
aioredis = "1.3.1"
sentry-sdk = "^1.5.7"
pyfixtures = "^1.0.0"
orjson = "^3.9.9"
sphinx-readable-theme = "^1.3.0"

[tool.poetry.scripts]
run-workflow = "virtool_workflow.cli:cli_main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-aiohttp = "^1.0.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
pre-commit = "^2.18.1"
sphinx = "^4.0.2"
syrupy = "^3.0.5"
pydantic-factories = "^1.17.3"
sphinx-autobuild = "^2021.3.14"
sphinx-toolbox = "^3.5.0"
sphinx-nameko-theme = "^0.0.3"
pytest-structlog = "^0.6"
biopython = "^1.81"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
exclude = [
    "docs",
    ".eggs",
    ".git",
    ".github",
    ".mypy_cache",
    ".pytest-cache",
    ".ruff_cache",
    "__pypackages__",
]
indent-width = 4
line-length = 88
target-version = "py310"

[tool.ruff.lint]
fixable = ["ALL", "I001"]

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
